Add files via upload
The script Smart_3d_registration is used to load data and applies some initial modifications such as white/black values (contrast adjustments). It calls the main registration fuction StackRegistration2018_5.

This function compares different 3d-registration algorithms included in MATLAB on a downscaled version of the stack to find which algo performs best, on a frame-to-frame basis. For each frame, the best algo is later applied to the full stack. 

The other functions are necessary to run the script, but are only little helper functions.
I used the bioformats-package for MATLAB to import the data (from: https://www.openmicroscopy.org/).